+++ src/main/java/com/aparapi/internal/jni/OpenCLJNI.java
@@ -32,7 +32,12 @@ public abstract class OpenCLJNI{
 
    protected native List<OpenCLPlatform> getPlatforms();
 
-   protected native OpenCLProgram createProgram(OpenCLDevice context, String openCLSource);
+   public OpenCLProgram createProgram(OpenCLDevice context, String openCLSource)
+   {
+      return this.createProgram(context, openCLSource, null);
+   }
+
+   protected native OpenCLProgram createProgram(OpenCLDevice context, String openCLSource, String binaryKey);
 
    protected native OpenCLKernel createKernelJNI(OpenCLProgram program, String kernelName, OpenCLArgDescriptor[] args);